Gate home skeletons and task fetch
